+++ src/app/app.module.ts
@@ -3,6 +3,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CartModule } from './cart/cart.module';
+import { CoreModule } from './core/core.module';
 import { FirstComponent } from './first/first.component';
 import { ProductsModule } from './products/products.module';
 
@@ -17,7 +18,7 @@ import { ProductsModule } from './products/products.module';
     ProductsModule,
     CartModule
   ],
-  providers: [],
+  providers: [CoreModule],
   bootstrap: [AppComponent]
 })
 export class AppModule { }