+++ src/app/cart/cart.module.ts
@@ -9,7 +9,7 @@ import { SharedModule } from '../shared/shared.module';
   imports: [
     BrowserModule,
     SharedModule
-    ],
+  ],
   exports: [CartListComponent],
   providers: [],
 })