+++ src/app/first/first.component.ts
@@ -1,4 +1,9 @@
-import { Component } from '@angular/core';
+import { Component, Inject, Optional } from '@angular/core';
+import { ConfigOptionsService } from '../core/services/config-options.service';
+import { ConstantsServiceToken, ConstantsServiceValue } from '../core/services/constants.service';
+import { GeneratorService } from '../core/services/generator';
+import { generatedString, GeneratorFactory } from '../core/services/generator.factory';
+import { localStorageToken } from '../core/services/local-storage.service';
 import { ProductCategories, ProductId } from '../products/models/product.model';
 
 @Component({
@@ -13,4 +18,13 @@ export class FirstComponent {
   category: ProductCategories = ProductCategories.White;
   isAvailable: boolean = false;
   ids: ProductId[] = ['1', '2', '3'];
+
+  constructor(
+     @Optional() private config: ConfigOptionsService,
+     @Optional() @Inject('ConstantsService') private constants: Record<string, unknown>,
+     @Optional() @Inject('GenerateString') private alfanumGenerator: (n: number) => string,
+     @Optional() private generator: GeneratorService,
+     @Optional() @Inject('localStorage') private localStorage: Storage,
+  ) {
+  }
 }