+++ src/app/products/components/product-list/product-list.component.ts
@@ -17,6 +17,6 @@ export class ProductListComponent {
   }
 
   addToCart(item: ProductModel): void {
-    this.cart.addItemToCart(item);
+    this.cart.addProduct(item);
   }
 }