+++ src/app/products/products.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { SharedModule } from '../shared/shared.module';
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { ProductComponent } from './components/product/product.component';
 
@@ -10,9 +11,9 @@ import { ProductComponent } from './components/product/product.component';
   ],
   imports: [
     BrowserModule,
+    SharedModule
   ],
   exports:[
-    ProductComponent,
     ProductListComponent,
   ],
   providers: [],